Add new module for networking
