+++ core/network/src/commonMain/kotlin/com/chesire/nekomp/core/network/ApiError.kt
@@ -1,0 +1,7 @@
+package com.chesire.nekomp.core.network
+
+data class ApiError(
+    val code: Int,
+    val reason: String = "",
+    val body: String
+) : Throwable()
